requirements.txt
Flask
openai
requests
Pillow